requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
gTTS==2.2.4
openai==0.19.0
fpdf==1.7.2
PyYAML==6.0
transformers==4.20.1
torch==1.11.0+cpu
torchvision==0.12.0+cpu
nltk==3.7
boto3==1.24.24
Flask==2.1.2
